+++ CHANGELOG.md
@@ -7,7 +7,11 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 
 ### Added
 
+- Add `ParserState::consumeWhiteSpaceWithComments()` method (#670)
+
 ### Changed
+
+- Redirect `ParserState::consumeWhiteSpace()` to `ParserState::consumeWhiteSpaceWithComments()` (#670)
 
 ### Deprecated